Refactored the ntfs accessors and allow remappings.
