+++ accessors/ntfs/fixtures/TestNTFSFilesystemAccessor.golden
@@ -1,0 +1,21 @@
+[
+ "/$AttrDef",
+ "/$BadClus",
+ "/$BadClus:$Bad",
+ "/$Bitmap",
+ "/$Boot",
+ "/$Extend",
+ "/$LogFile",
+ "/$MFT",
+ "/$MFTMirr",
+ "/$RECYCLE.BIN",
+ "/$Secure",
+ "/$Secure:$SDS",
+ "/$UpCase",
+ "/$UpCase:$Info",
+ "/$Volume",
+ "/.",
+ "/Folder A",
+ "/System Volume Information",
+ "/ones.bin"
+]
