+++ accessors/ntfs/fixtures/TestNTFSFilesystemAccessorRemapping.golden
@@ -1,0 +1,4 @@
+[
+ "\\\\.\\C:\\$MFT",
+ "\\\\.\\D:\\$MFT"
+]
